infra/ml/requirements.txt
pmdarima
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv